Move currency-related functionality into Casino/Currency
