+++ _BotNameTest/_BotNameTest.csproj
@@ -39,9 +39,9 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="CasinoLogic\CasinoClaimTest.cs" />
+    <Compile Include="Casino\Currency\CasinoClaimTest.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
-    <Compile Include="CasinoLogic\CasinoShopTest.cs" />
+    <Compile Include="Casino\CasinoShopTest.cs" />
   </ItemGroup>
   <ItemGroup>
     <ProjectReference Include="..\_BotName\_BotName.csproj">
